kafka-debezium/requirements.txt
kafka-python-ng
apache-airflow-client
python-dotenv